Change default to 3.9.1 and put 3.8.0 as alternative
