+++ zookeeper-server/pom.xml
@@ -13,8 +13,8 @@
   <version>8-SNAPSHOT</version>
   <modules>
     <module>zookeeper-server-common</module>
+    <module>zookeeper-server-3.8.0</module>
     <module>zookeeper-server</module>
-    <module>zookeeper-server-3.9.1</module>
   </modules>
   <dependencies>
     <dependency>
