+++ zookeeper-server/zookeeper-server-3.8.0/src/main/java/org/apache/zookeeper/server/ZooKeeperServer.java
@@ -1,3 +1,4 @@
+// Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 /*
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
@@ -18,33 +19,8 @@
 
 package org.apache.zookeeper.server;
 
-import java.io.BufferedInputStream;
-import java.io.ByteArrayOutputStream;
-import java.io.File;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.PrintWriter;
-import java.nio.ByteBuffer;
-import java.util.ArrayDeque;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Deque;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Properties;
-import java.util.Random;
-import java.util.Set;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.atomic.AtomicInteger;
-import java.util.concurrent.atomic.AtomicLong;
-import java.util.function.BiConsumer;
-import java.util.zip.Adler32;
-import java.util.zip.CheckedInputStream;
-import javax.security.sasl.SaslException;
 import org.apache.jute.BinaryInputArchive;
 import org.apache.jute.BinaryOutputArchive;
-import org.apache.jute.InputArchive;
 import org.apache.jute.Record;
 import org.apache.zookeeper.Environment;
 import org.apache.zookeeper.KeeperException;
@@ -96,6 +72,27 @@ import org.apache.zookeeper.util.ServiceUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import javax.security.sasl.SaslException;
+import java.io.ByteArrayOutputStream;
+import java.io.File;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.PrintWriter;
+import java.nio.ByteBuffer;
+import java.util.ArrayDeque;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Deque;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Properties;
+import java.util.Random;
+import java.util.Set;
+import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
+import java.util.function.BiConsumer;
+
 /**
  * This class implements a simple standalone ZooKeeperServer. It sets up the
  * following chain of RequestProcessors to process requests:
@@ -114,7 +111,7 @@ public class ZooKeeperServer implements SessionExpirer, ServerStats.Provider {
 
     // When enabled, will check ACL constraints appertained to the requests first,
     // before sending the requests to the quorum.
-    static boolean enableEagerACLCheck;
+    static final boolean enableEagerACLCheck;
 
     static final boolean skipACL;
 
@@ -126,14 +123,10 @@ public class ZooKeeperServer implements SessionExpirer, ServerStats.Provider {
     public static final String ZOOKEEPER_DIGEST_ENABLED = "zookeeper.digest.enabled";
     private static boolean digestEnabled;
 
-    public static final String ZOOKEEPER_SERIALIZE_LAST_PROCESSED_ZXID_ENABLED = "zookeeper.serializeLastProcessedZxid.enabled";
-    private static boolean serializeLastProcessedZxidEnabled;
-
     // Add a enable/disable option for now, we should remove this one when
     // this feature is confirmed to be stable
     public static final String CLOSE_SESSION_TXN_ENABLED = "zookeeper.closeSessionTxn.enabled";
     private static boolean closeSessionTxnEnabled = true;
-    private volatile CountDownLatch restoreLatch;
 
     static {
         LOG = LoggerFactory.getLogger(ZooKeeperServer.class);
@@ -163,20 +156,6 @@ public class ZooKeeperServer implements SessionExpirer, ServerStats.Provider {
         closeSessionTxnEnabled = Boolean.parseBoolean(
                 System.getProperty(CLOSE_SESSION_TXN_ENABLED, "true"));
         LOG.info("{} = {}", CLOSE_SESSION_TXN_ENABLED, closeSessionTxnEnabled);
-
-        setSerializeLastProcessedZxidEnabled(Boolean.parseBoolean(
-                System.getProperty(ZOOKEEPER_SERIALIZE_LAST_PROCESSED_ZXID_ENABLED, "true")));
-    }
-
-    // @VisibleForTesting
-    public static boolean isEnableEagerACLCheck() {
-        return enableEagerACLCheck;
-    }
-
-    // @VisibleForTesting
-    public static void setEnableEagerACLCheck(boolean enabled) {
-        ZooKeeperServer.enableEagerACLCheck = enabled;
-        LOG.info("Update {} to {}", ENABLE_EAGER_ACL_CHECK, enabled);
     }
 
     public static boolean isCloseSessionTxnEnabled() {
@@ -240,7 +219,7 @@ public class ZooKeeperServer implements SessionExpirer, ServerStats.Provider {
     private final AtomicInteger requestsInProcess = new AtomicInteger(0);
     final Deque<ChangeRecord> outstandingChanges = new ArrayDeque<>();
     // this data structure must be accessed under the outstandingChanges lock
-    final Map<String, ChangeRecord> outstandingChangesForPath = new HashMap<>();
+    final Map<String, ChangeRecord> outstandingChangesForPath = new HashMap<String, ChangeRecord>();
 
     protected ServerCnxnFactory serverCnxnFactory;
     protected ServerCnxnFactory secureServerCnxnFactory;
@@ -287,7 +266,7 @@ public class ZooKeeperServer implements SessionExpirer, ServerStats.Provider {
     }
 
     // Connection throttling
-    private final BlueThrottle connThrottle = new BlueThrottle();
+    private BlueThrottle connThrottle = new BlueThrottle();
 
     private RequestThrottler requestThrottler;
     public static final String SNAP_COUNT = "zookeeper.snapCount";
@@ -298,17 +277,17 @@ public class ZooKeeperServer implements SessionExpirer, ServerStats.Provider {
      * too many large requests such that the JVM runs out of usable heap and
      * ultimately crashes.
      *
-     * The limit is enforced by the {@link #checkRequestSizeWhenReceivingMessage(int)}
+     * The limit is enforced by the {@link checkRequestSize(int, boolean)}
      * method which is called by the connection layer ({@link NIOServerCnxn},
      * {@link NettyServerCnxn}) before allocating a byte buffer and pulling
      * data off the TCP socket. The limit is then checked again by the
-     * ZooKeeper server in {@link #processPacket(ServerCnxn, RequestHeader, RequestRecord)} which
-     * also atomically updates {@link #currentLargeRequestBytes}. The request is
+     * ZooKeeper server in {@link processPacket(ServerCnxn, ByteBuffer)} which
+     * also atomically updates {@link currentLargeRequestBytes}. The request is
      * then marked as a large request, with the request size stored in the Request
-     * object so that it can later be decremented from {@link #currentLargeRequestBytes}.
+     * object so that it can later be decremented from {@link currentLargeRequestsBytes}.
      *
      * When a request is completed or dropped, the relevant code path calls the
-     * {@link #requestFinished(Request)} method which performs the decrement if
+     * {@link requestFinished(Request)} method which performs the decrement if
      * needed.
      */
     private volatile int largeRequestMaxBytes = 100 * 1024 * 1024;
@@ -321,7 +300,7 @@ public class ZooKeeperServer implements SessionExpirer, ServerStats.Provider {
 
     private final AtomicInteger currentLargeRequestBytes = new AtomicInteger(0);
 
-    private final AuthenticationHelper authHelper = new AuthenticationHelper();
+    private AuthenticationHelper authHelper;
 
     void removeCnxn(ServerCnxn cnxn) {
         zkDb.removeCnxn(cnxn);
@@ -337,6 +316,7 @@ public class ZooKeeperServer implements SessionExpirer, ServerStats.Provider {
         listener = new ZooKeeperServerListenerImpl(this);
         serverStats = new ServerStats(this);
         this.requestPathMetricsCollector = new RequestPathMetricsCollector();
+        this.authHelper = new AuthenticationHelper();
     }
 
     /**
@@ -378,8 +358,10 @@ public class ZooKeeperServer implements SessionExpirer, ServerStats.Provider {
 
         this.initLargeRequestThrottlingSettings();
 
+        this.authHelper = new AuthenticationHelper();
+
         LOG.info(
-            "Created server with"
+            "Created patched server with"
                 + " tickTime {} ms"
                 + " minSessionTimeout {} ms"
                 + " maxSessionTimeout {} ms"
@@ -544,100 +526,23 @@ public class ZooKeeperServer implements SessionExpirer, ServerStats.Provider {
         takeSnapshot();
     }
 
-    public File takeSnapshot() throws IOException {
-        return takeSnapshot(false);
+    public void takeSnapshot() {
+        takeSnapshot(false);
     }
 
-    public File takeSnapshot(boolean syncSnap) throws IOException {
-        return takeSnapshot(syncSnap, true, false);
-    }
-
-    /**
-     * Takes a snapshot on the server.
-     *
-     * @param syncSnap syncSnap sync the snapshot immediately after write
-     * @param isSevere if true system exist, otherwise throw IOException
-     * @param fastForwardFromEdits whether fast forward database to the latest recorded transactions
-     *
-     * @return file snapshot file object
-     * @throws IOException
-     */
-    public synchronized File takeSnapshot(boolean syncSnap, boolean isSevere, boolean fastForwardFromEdits) throws IOException {
+    public void takeSnapshot(boolean syncSnap) {
         long start = Time.currentElapsedTime();
-        File snapFile = null;
         try {
-            if (fastForwardFromEdits) {
-                zkDb.fastForwardDataBase();
-            }
-            snapFile = txnLogFactory.save(zkDb.getDataTree(), zkDb.getSessionWithTimeOuts(), syncSnap);
+            txnLogFactory.save(zkDb.getDataTree(), zkDb.getSessionWithTimeOuts(), syncSnap);
         } catch (IOException e) {
-            if (isSevere) {
-                LOG.error("Severe unrecoverable error, exiting", e);
-                // This is a severe error that we cannot recover from,
-                // so we need to exit
-                ServiceUtils.requestSystemExit(ExitCode.TXNLOG_ERROR_TAKING_SNAPSHOT.getValue());
-            } else {
-                throw e;
-            }
+            LOG.error("Severe unrecoverable error, exiting", e);
+            // This is a severe error that we cannot recover from,
+            // so we need to exit
+            ServiceUtils.requestSystemExit(ExitCode.TXNLOG_ERROR_TAKING_SNAPSHOT.getValue());
         }
         long elapsed = Time.currentElapsedTime() - start;
         LOG.info("Snapshot taken in {} ms", elapsed);
         ServerMetrics.getMetrics().SNAPSHOT_TIME.add(elapsed);
-        return snapFile;
-    }
-
-    /**
-     * Restores database from a snapshot. It is used by the restore admin server command.
-     *
-     * @param inputStream input stream of snapshot
-     * @return last processed zxid
-     */
-    public synchronized long restoreFromSnapshot(final InputStream inputStream) throws IOException {
-        if (inputStream == null) {
-            throw new IllegalArgumentException("InputStream can not be null when restoring from snapshot");
-        }
-
-        long start = Time.currentElapsedTime();
-        LOG.info("Before restore database. lastProcessedZxid={}, nodeCount={}，sessionCount={}",
-            getZKDatabase().getDataTreeLastProcessedZxid(),
-            getZKDatabase().dataTree.getNodeCount(),
-            getZKDatabase().getSessionCount());
-
-        // restore to a new zkDatabase
-        final ZKDatabase newZKDatabase = new ZKDatabase(this.txnLogFactory);
-        final CheckedInputStream cis = new CheckedInputStream(new BufferedInputStream(inputStream), new Adler32());
-        final InputArchive ia = BinaryInputArchive.getArchive(cis);
-        newZKDatabase.deserializeSnapshot(ia, cis);
-        LOG.info("Restored to a new database. lastProcessedZxid={}, nodeCount={}, sessionCount={}",
-            newZKDatabase.getDataTreeLastProcessedZxid(),
-            newZKDatabase.dataTree.getNodeCount(),
-            newZKDatabase.getSessionCount());
-
-        // create a CountDownLatch
-        restoreLatch = new CountDownLatch(1);
-
-        try {
-            // set to the new zkDatabase
-            setZKDatabase(newZKDatabase);
-
-            // re-create SessionTrack
-            createSessionTracker();
-        } finally {
-            // unblock request submission
-            restoreLatch.countDown();
-            restoreLatch = null;
-        }
-
-        LOG.info("After restore database. lastProcessedZxid={}, nodeCount={}, sessionCount={}",
-                getZKDatabase().getDataTreeLastProcessedZxid(),
-                getZKDatabase().dataTree.getNodeCount(),
-                getZKDatabase().getSessionCount());
-
-        long elapsed = Time.currentElapsedTime() - start;
-        LOG.info("Restore taken in {} ms", elapsed);
-        ServerMetrics.getMetrics().RESTORE_TIME.add(elapsed);
-
-        return getLastProcessedZxid();
     }
 
     public boolean shouldForceWriteInitialSnapshotAfterLeaderElection() {
@@ -830,12 +735,9 @@ public class ZooKeeperServer implements SessionExpirer, ServerStats.Provider {
     }
 
     protected void startRequestThrottler() {
-        requestThrottler = createRequestThrottler();
+        requestThrottler = new RequestThrottler(this);
         requestThrottler.start();
-    }
 
-    protected RequestThrottler createRequestThrottler() {
-        return new RequestThrottler(this);
     }
 
     protected void setupRequestProcessors() {
@@ -881,10 +783,7 @@ public class ZooKeeperServer implements SessionExpirer, ServerStats.Provider {
      * error events, e.g., SyncRequestProcessor not being able to write a txn to
      * disk.</li>
      * <li>During shutdown the server sets the state to SHUTDOWN, which
-     * corresponds to the server not running.</li>
-     *
-     * <li>During maintenance (e.g. restore) the server sets the state to MAINTENANCE
-     * </li></ul>
+     * corresponds to the server not running.</li></ul>
      *
      * @param state new server state.
      */
@@ -1099,9 +998,10 @@ public class ZooKeeperServer implements SessionExpirer, ServerStats.Provider {
         long sessionId = sessionTracker.createSession(timeout);
         Random r = new Random(sessionId ^ superSecret);
         r.nextBytes(passwd);
-        CreateSessionTxn txn = new CreateSessionTxn(timeout);
+        ByteBuffer to = ByteBuffer.allocate(4);
+        to.putInt(timeout);
         cnxn.setSessionId(sessionId);
-        Request si = new Request(cnxn, sessionId, 0, OpCode.createSession, RequestRecord.fromRecord(txn), null);
+        Request si = new Request(cnxn, sessionId, 0, OpCode.createSession, to, null);
         submitRequest(si);
         return sessionId;
     }
@@ -1159,12 +1059,14 @@ public class ZooKeeperServer implements SessionExpirer, ServerStats.Provider {
                 valid ? cnxn.getSessionTimeout() : 0,
                 valid ? cnxn.getSessionId() : 0, // send 0 if session is no
                 // longer valid
-                valid ? generatePasswd(cnxn.getSessionId()) : new byte[16],
-                this instanceof ReadOnlyZooKeeperServer);
+                valid ? generatePasswd(cnxn.getSessionId()) : new byte[16]);
             ByteArrayOutputStream baos = new ByteArrayOutputStream();
             BinaryOutputArchive bos = BinaryOutputArchive.getArchive(baos);
             bos.writeInt(-1, "len");
             rsp.serialize(bos, "connect");
+            if (!cnxn.isOldClient) {
+                bos.writeBool(this instanceof ReadOnlyZooKeeperServer, "readOnly");
+            }
             baos.close();
             ByteBuffer bb = ByteBuffer.wrap(baos.toByteArray());
             bb.putInt(bb.remaining() - 4).rewind();
@@ -1211,14 +1113,6 @@ public class ZooKeeperServer implements SessionExpirer, ServerStats.Provider {
     }
 
     public void submitRequest(Request si) {
-        if (restoreLatch != null) {
-            try {
-                LOG.info("Blocking request submission while restore is in progress");
-                restoreLatch.await();
-            } catch (final InterruptedException e) {
-                LOG.warn("Unexpected interruption", e);
-            }
-        }
         enqueueRequest(si);
     }
 
@@ -1468,13 +1362,18 @@ public class ZooKeeperServer implements SessionExpirer, ServerStats.Provider {
         return connThrottle.getDropChance();
     }
 
-    public void processConnectRequest(ServerCnxn cnxn, ConnectRequest request) throws IOException, ClientCnxnLimitException {
+    public void processConnectRequest(ServerCnxn cnxn, ByteBuffer incomingBuffer)
+        throws IOException, ClientCnxnLimitException {
+
+        BinaryInputArchive bia = BinaryInputArchive.getArchive(new ByteBufferInputStream(incomingBuffer));
+        ConnectRequest connReq = new ConnectRequest();
+        connReq.deserialize(bia, "connect");
         LOG.debug(
             "Session establishment request from client {} client's lastZxid is 0x{}",
             cnxn.getRemoteSocketAddress(),
-            Long.toHexString(request.getLastZxidSeen()));
+            Long.toHexString(connReq.getLastZxidSeen()));
 
-        long sessionId = request.getSessionId();
+        long sessionId = connReq.getSessionId();
         int tokensNeeded = 1;
         if (connThrottle.isConnectionWeightEnabled()) {
             if (sessionId == 0) {
@@ -1492,24 +1391,30 @@ public class ZooKeeperServer implements SessionExpirer, ServerStats.Provider {
             throw new ClientCnxnLimitException();
         }
         ServerMetrics.getMetrics().CONNECTION_TOKEN_DEFICIT.add(connThrottle.getDeficit());
+
         ServerMetrics.getMetrics().CONNECTION_REQUEST_COUNT.add(1);
 
-        if (!cnxn.protocolManager.isReadonlyAvailable()) {
+        boolean readOnly = false;
+        try {
+            readOnly = bia.readBool("readOnly");
+            cnxn.isOldClient = false;
+        } catch (IOException e) {
+            // this is ok -- just a packet from an old client which
+            // doesn't contain readOnly field
             LOG.warn(
                 "Connection request from old client {}; will be dropped if server is in r-o mode",
                 cnxn.getRemoteSocketAddress());
         }
-
-        if (!request.getReadOnly() && this instanceof ReadOnlyZooKeeperServer) {
+        if (!readOnly && this instanceof ReadOnlyZooKeeperServer) {
             String msg = "Refusing session request for not-read-only client " + cnxn.getRemoteSocketAddress();
             LOG.info(msg);
             throw new CloseRequestException(msg, ServerCnxn.DisconnectReason.NOT_READ_ONLY_CLIENT);
         }
-        if (request.getLastZxidSeen() > zkDb.dataTree.lastProcessedZxid) {
+        if (connReq.getLastZxidSeen() > zkDb.dataTree.lastProcessedZxid) {
             String msg = "Refusing session request for client "
                          + cnxn.getRemoteSocketAddress()
                          + " as it has seen zxid 0x"
-                         + Long.toHexString(request.getLastZxidSeen())
+                         + Long.toHexString(connReq.getLastZxidSeen())
                          + " our last zxid is 0x"
                          + Long.toHexString(getZKDatabase().getDataTreeLastProcessedZxid())
                          + " client must try another server";
@@ -1517,8 +1422,8 @@ public class ZooKeeperServer implements SessionExpirer, ServerStats.Provider {
             LOG.info(msg);
             throw new CloseRequestException(msg, ServerCnxn.DisconnectReason.CLIENT_ZXID_AHEAD);
         }
-        int sessionTimeout = request.getTimeOut();
-        byte[] passwd = request.getPasswd();
+        int sessionTimeout = connReq.getTimeOut();
+        byte[] passwd = connReq.getPasswd();
         int minSessionTimeout = getMinSessionTimeout();
         if (sessionTimeout < minSessionTimeout) {
             sessionTimeout = minSessionTimeout;
@@ -1536,16 +1441,16 @@ public class ZooKeeperServer implements SessionExpirer, ServerStats.Provider {
             LOG.debug(
                 "Client attempting to establish new session: session = 0x{}, zxid = 0x{}, timeout = {}, address = {}",
                 Long.toHexString(id),
-                Long.toHexString(request.getLastZxidSeen()),
-                request.getTimeOut(),
+                Long.toHexString(connReq.getLastZxidSeen()),
+                connReq.getTimeOut(),
                 cnxn.getRemoteSocketAddress());
         } else {
             validateSession(cnxn, sessionId);
             LOG.debug(
                 "Client attempting to renew session: session = 0x{}, zxid = 0x{}, timeout = {}, address = {}",
                 Long.toHexString(sessionId),
-                Long.toHexString(request.getLastZxidSeen()),
-                request.getTimeOut(),
+                Long.toHexString(connReq.getLastZxidSeen()),
+                connReq.getTimeOut(),
                 cnxn.getRemoteSocketAddress());
             if (serverCnxnFactory != null) {
                 serverCnxnFactory.closeSession(sessionId, ServerCnxn.DisconnectReason.CLIENT_RECONNECT);
@@ -1685,7 +1590,13 @@ public class ZooKeeperServer implements SessionExpirer, ServerStats.Provider {
         }
     }
 
-    public void processPacket(ServerCnxn cnxn, RequestHeader h, RequestRecord request) throws IOException {
+    public void processPacket(ServerCnxn cnxn, ByteBuffer incomingBuffer) throws IOException {
+        // We have the request, now process and setup for next
+        InputStream bais = new ByteBufferInputStream(incomingBuffer);
+        BinaryInputArchive bia = BinaryInputArchive.getArchive(bais);
+        RequestHeader h = new RequestHeader();
+        h.deserialize(bia, "header");
+
         // Need to increase the outstanding request count first, otherwise
         // there might be a race condition that it enabled recv after
         // processing request and then disabled when check throttling.
@@ -1697,12 +1608,17 @@ public class ZooKeeperServer implements SessionExpirer, ServerStats.Provider {
         // in cnxn, since it will close the cnxn anyway.
         cnxn.incrOutstandingAndCheckThrottle(h);
 
+        // Through the magic of byte buffers, txn will not be
+        // pointing
+        // to the start of the txn
+        incomingBuffer = incomingBuffer.slice();
         if (h.getType() == OpCode.auth) {
             LOG.info("got auth packet {}", cnxn.getRemoteSocketAddress());
-            AuthPacket authPacket = request.readRecord(AuthPacket::new);
+            AuthPacket authPacket = new AuthPacket();
+            ByteBufferInputStream.byteBuffer2Record(incomingBuffer, authPacket);
             String scheme = authPacket.getScheme();
             ServerAuthenticationProvider ap = ProviderRegistry.getServerProvider(scheme);
-            Code authReturn = KeeperException.Code.AUTHFAILED;
+            Code authReturn = Code.AUTHFAILED;
             if (ap != null) {
                 try {
                     // handleAuthentication may close the connection, to allow the client to choose
@@ -1712,14 +1628,14 @@ public class ZooKeeperServer implements SessionExpirer, ServerStats.Provider {
                         authPacket.getAuth());
                 } catch (RuntimeException e) {
                     LOG.warn("Caught runtime exception from AuthenticationProvider: {}", scheme, e);
-                    authReturn = KeeperException.Code.AUTHFAILED;
+                    authReturn = Code.AUTHFAILED;
                 }
             }
-            if (authReturn == KeeperException.Code.OK) {
+            if (authReturn == Code.OK) {
                 LOG.info("Session 0x{}: auth success for scheme {} and address {}",
                         Long.toHexString(cnxn.getSessionId()), scheme,
                         cnxn.getRemoteSocketAddress());
-                ReplyHeader rh = new ReplyHeader(h.getXid(), 0, KeeperException.Code.OK.intValue());
+                ReplyHeader rh = new ReplyHeader(h.getXid(), 0, Code.OK.intValue());
                 cnxn.sendResponse(rh, null, null);
             } else {
                 if (ap == null) {
@@ -1731,7 +1647,7 @@ public class ZooKeeperServer implements SessionExpirer, ServerStats.Provider {
                     LOG.warn("Authentication failed for scheme: {}", scheme);
                 }
                 // send a response...
-                ReplyHeader rh = new ReplyHeader(h.getXid(), 0, KeeperException.Code.AUTHFAILED.intValue());
+                ReplyHeader rh = new ReplyHeader(h.getXid(), 0, Code.AUTHFAILED.intValue());
                 cnxn.sendResponse(rh, null, null);
                 // ... and close connection
                 cnxn.sendBuffer(ServerCnxnFactory.closeConn);
@@ -1739,15 +1655,15 @@ public class ZooKeeperServer implements SessionExpirer, ServerStats.Provider {
             }
             return;
         } else if (h.getType() == OpCode.sasl) {
-            processSasl(request, cnxn, h);
+            processSasl(incomingBuffer, cnxn, h);
         } else {
             if (!authHelper.enforceAuthentication(cnxn, h.getXid())) {
                 // Authentication enforcement is failed
                 // Already sent response to user about failure and closed the session, lets return
                 return;
             } else {
-                Request si = new Request(cnxn, cnxn.getSessionId(), h.getXid(), h.getType(), request, cnxn.getAuthInfo());
-                int length = request.limit();
+                Request si = new Request(cnxn, cnxn.getSessionId(), h.getXid(), h.getType(), incomingBuffer, cnxn.getAuthInfo());
+                int length = incomingBuffer.limit();
                 if (isLargeRequest(length)) {
                     // checkRequestSize will throw IOException if request is rejected
                     checkRequestSizeWhenMessageReceived(length);
@@ -1785,9 +1701,10 @@ public class ZooKeeperServer implements SessionExpirer, ServerStats.Provider {
         return Boolean.getBoolean(ALLOW_SASL_FAILED_CLIENTS);
     }
 
-    private void processSasl(RequestRecord request, ServerCnxn cnxn, RequestHeader requestHeader) throws IOException {
+    private void processSasl(ByteBuffer incomingBuffer, ServerCnxn cnxn, RequestHeader requestHeader) throws IOException {
         LOG.debug("Responding to client SASL token.");
-        GetSASLRequest clientTokenRecord = request.readRecord(GetSASLRequest::new);
+        GetSASLRequest clientTokenRecord = new GetSASLRequest();
+        ByteBufferInputStream.byteBuffer2Record(incomingBuffer, clientTokenRecord);
         byte[] clientToken = clientTokenRecord.getToken();
         LOG.debug("Size of client SASL token: {}", clientToken.length);
         byte[] responseToken = null;
@@ -2062,7 +1979,7 @@ public class ZooKeeperServer implements SessionExpirer, ServerStats.Provider {
 
     /**
      * Grant or deny authorization to an operation on a node as a function of:
-     * @param cnxn :    the server connection or null for admin server commands
+     * @param cnxn :    the server connection
      * @param acl :     set of ACLs for the node
      * @param perm :    the permission that the client is requesting
      * @param ids :     the credentials supplied by the client
@@ -2235,15 +2152,6 @@ public class ZooKeeperServer implements SessionExpirer, ServerStats.Provider {
         ZooKeeperServer.digestEnabled = digestEnabled;
     }
 
-    public static boolean isSerializeLastProcessedZxidEnabled() {
-        return serializeLastProcessedZxidEnabled;
-    }
-
-    public static void setSerializeLastProcessedZxidEnabled(boolean serializeLastZxidEnabled) {
-        serializeLastProcessedZxidEnabled = serializeLastZxidEnabled;
-        LOG.info("{} = {}", ZOOKEEPER_SERIALIZE_LAST_PROCESSED_ZXID_ENABLED, serializeLastZxidEnabled);
-    }
-
     /**
      * Trim a path to get the immediate predecessor.
      *
@@ -2267,8 +2175,8 @@ public class ZooKeeperServer implements SessionExpirer, ServerStats.Provider {
         switch (request.type) {
         case OpCode.create:
         case OpCode.create2: {
-            CreateRequest req = request.readRequestRecordNoException(CreateRequest::new);
-            if (req != null) {
+            CreateRequest req = new CreateRequest();
+            if (buffer2Record(request.request, req)) {
                 mustCheckACL = true;
                 acl = req.getAcl();
                 path = parentPath(req.getPath());
@@ -2276,22 +2184,22 @@ public class ZooKeeperServer implements SessionExpirer, ServerStats.Provider {
             break;
         }
         case OpCode.delete: {
-            DeleteRequest req = request.readRequestRecordNoException(DeleteRequest::new);
-            if (req != null) {
+            DeleteRequest req = new DeleteRequest();
+            if (buffer2Record(request.request, req)) {
                 path = parentPath(req.getPath());
             }
             break;
         }
         case OpCode.setData: {
-            SetDataRequest req = request.readRequestRecordNoException(SetDataRequest::new);
-            if (req != null) {
+            SetDataRequest req = new SetDataRequest();
+            if (buffer2Record(request.request, req)) {
                 path = req.getPath();
             }
             break;
         }
         case OpCode.setACL: {
-            SetACLRequest req = request.readRequestRecordNoException(SetACLRequest::new);
-            if (req != null) {
+            SetACLRequest req = new SetACLRequest();
+            if (buffer2Record(request.request, req)) {
                 mustCheckACL = true;
                 acl = req.getAcl();
                 path = req.getPath();
@@ -2348,7 +2256,7 @@ public class ZooKeeperServer implements SessionExpirer, ServerStats.Provider {
             return true;
         }
 
-        err = KeeperException.Code.OK.intValue();
+        err = Code.OK.intValue();
 
         try {
             pathToCheck = effectiveACLPath(request);
@@ -2369,7 +2277,7 @@ public class ZooKeeperServer implements SessionExpirer, ServerStats.Provider {
             LOG.error("Uncaught exception in authWriteRequest with: ", t);
             throw t;
         } finally {
-            if (err != KeeperException.Code.OK.intValue()) {
+            if (err != Code.OK.intValue()) {
                 /*  This request has a bad ACL, so we are dismissing it early. */
                 decInProcess();
                 ReplyHeader rh = new ReplyHeader(request.cxid, 0, err);
@@ -2381,7 +2289,19 @@ public class ZooKeeperServer implements SessionExpirer, ServerStats.Provider {
             }
         }
 
-        return err == KeeperException.Code.OK.intValue();
+        return err == Code.OK.intValue();
+    }
+
+    private boolean buffer2Record(ByteBuffer request, Record record) {
+        boolean rv = false;
+        try {
+            ByteBufferInputStream.byteBuffer2Record(request, record);
+            request.rewind();
+            rv = true;
+        } catch (IOException ex) {
+        }
+
+        return rv;
     }
 
     public int getOutstandingHandshakeNum() {
