+++ zookeeper-server/zookeeper-server-3.8.0/src/main/java/org/apache/zookeeper/server/quorum/Learner.java
@@ -1,3 +1,4 @@
+// Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 /*
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
@@ -133,7 +134,7 @@ public class Learner {
         LOG.info("{} = {}", LEARNER_CLOSE_SOCKET_ASYNC, closeSocketAsync);
     }
 
-    final ConcurrentHashMap<Long, ServerCnxn> pendingRevalidations = new ConcurrentHashMap<>();
+    final ConcurrentHashMap<Long, ServerCnxn> pendingRevalidations = new ConcurrentHashMap<Long, ServerCnxn>();
 
     public int getPendingRevalidationsCount() {
         return pendingRevalidations.size();
@@ -254,9 +255,13 @@ public class Learner {
         oa.writeLong(request.sessionId);
         oa.writeInt(request.cxid);
         oa.writeInt(request.type);
-        byte[] payload = request.readRequestBytes();
-        if (payload != null) {
-            oa.write(payload);
+        if (request.request != null) {
+            request.request.rewind();
+            int len = request.request.remaining();
+            byte[] b = new byte[len];
+            request.request.get(b);
+            request.request.rewind();
+            oa.write(b);
         }
         oa.close();
         QuorumPacket qp = new QuorumPacket(Leader.REQUEST, -1, baos.toByteArray(), request.authInfo);
@@ -491,7 +496,7 @@ public class Learner {
         /*
          * Add sid to payload
          */
-        LearnerInfo li = new LearnerInfo(self.getMyId(), 0x10000, self.getQuorumVerifier().getVersion());
+        LearnerInfo li = new LearnerInfo(self.getId(), 0x10000, self.getQuorumVerifier().getVersion());
         ByteArrayOutputStream bsid = new ByteArrayOutputStream();
         BinaryOutputArchive boa = BinaryOutputArchive.getArchive(bsid);
         boa.writeRecord(li, "LearnerInfo");
@@ -543,140 +548,36 @@ public class Learner {
      * @throws InterruptedException
      */
     protected void syncWithLeader(long newLeaderZxid) throws Exception {
+        QuorumPacket ack = new QuorumPacket(Leader.ACK, 0, null, null);
+        QuorumPacket qp = new QuorumPacket();
         long newEpoch = ZxidUtils.getEpochFromZxid(newLeaderZxid);
+
         QuorumVerifier newLeaderQV = null;
 
-        class SyncHelper {
-
-            // In the DIFF case we don't need to do a snapshot because the transactions will sync on top of any existing snapshot.
-            // For SNAP and TRUNC the snapshot is needed to save that history.
-            boolean willSnapshot = true;
-            boolean syncSnapshot = false;
-
-            // PROPOSALs received during sync, for matching up with COMMITs.
-            Deque<PacketInFlight> proposals = new ArrayDeque<>();
-
-            // PROPOSALs we delay forwarding to the ZK server until sync is done.
-            Deque<PacketInFlight> delayedProposals = new ArrayDeque<>();
-
-            // COMMITs we delay forwarding to the ZK server until sync is done.
-            Deque<Long> delayedCommits = new ArrayDeque<>();
-
-            void syncSnapshot() {
-                syncSnapshot = true;
-            }
-
-            void noSnapshot() {
-                willSnapshot = false;
-            }
-
-            void propose(PacketInFlight pif) {
-                proposals.add(pif);
-                delayedProposals.add(pif);
-            }
-
-            PacketInFlight nextProposal() {
-                return proposals.peekFirst();
-            }
-
-            void commit() {
-                PacketInFlight packet = proposals.remove();
-                if (willSnapshot) {
-                    zk.processTxn(packet.hdr, packet.rec);
-                    delayedProposals.remove();
-                } else {
-                    delayedCommits.add(packet.hdr.getZxid());
-                }
-            }
-
-            void writeState() throws IOException, InterruptedException {
-                // Ensure all received transaction PROPOSALs are written before we ACK the NEWLEADER,
-                // since this allows the leader to apply those transactions to its served state:
-                if (willSnapshot) {
-                    zk.takeSnapshot(syncSnapshot); // either, the snapshot contains the transactions,
-                    willSnapshot = false;          // but anything after this needs to go to the transaction log; or
-                }
-
-                self.setCurrentEpoch(newEpoch);
-                sock.setSoTimeout(self.tickTime * self.syncLimit);
-                self.setSyncMode(QuorumPeer.SyncMode.NONE);
-                zk.startupWithoutServing();
-
-                // if we're a follower, we need to ensure the transactions are safely logged before ACK'ing.
-                if (zk instanceof FollowerZooKeeperServer) {
-                    FollowerZooKeeperServer fzk = (FollowerZooKeeperServer) zk;
-                    // The leader expects the NEWLEADER ACK to precede all the PROPOSAL ACKs, so we only write them first.
-                    fzk.syncProcessor.setDelayForwarding(true);
-                    for (PacketInFlight p : delayedProposals) {
-                        fzk.logRequest(p.hdr, p.rec, p.digest);
-                    }
-                    delayedProposals.clear();
-                    fzk.syncProcessor.syncFlush();
-                }
-            }
-
-            void flushAcks() throws InterruptedException {
-                if (zk instanceof FollowerZooKeeperServer) {
-                    // The NEWLEADER is ACK'ed, and we can now ACK the PROPOSALs we wrote in writeState.
-                    FollowerZooKeeperServer fzk = (FollowerZooKeeperServer) zk;
-                    fzk.syncProcessor.setDelayForwarding(false);
-                    fzk.syncProcessor.syncFlush(); // Ensure these are all ACK'ed before the UPTODATE ACK.
-                }
-            }
-
-            void applyDelayedPackets() {
-                // Any delayed packets must now be applied: all PROPOSALs first, then any COMMITs.
-                if (zk instanceof FollowerZooKeeperServer) {
-                    FollowerZooKeeperServer fzk = (FollowerZooKeeperServer) zk;
-                    for (PacketInFlight p : delayedProposals) {
-                        fzk.logRequest(p.hdr, p.rec, p.digest);
-                    }
-                    for (Long zxid : delayedCommits) {
-                        fzk.commit(zxid);
-                    }
-                } else if (zk instanceof ObserverZooKeeperServer) {
-                    ObserverZooKeeperServer ozk = (ObserverZooKeeperServer) zk;
-                    for (PacketInFlight p : delayedProposals) {
-                        Long zxid = delayedCommits.peekFirst();
-                        if (p.hdr.getZxid() != zxid) {
-                            // log warning message if there is no matching commit
-                            // old leader send outstanding proposal to observer
-                            LOG.warn(
-                                    "Committing 0x{}, but next proposal is 0x{}",
-                                    Long.toHexString(zxid),
-                                    Long.toHexString(p.hdr.getZxid()));
-                            continue;
-                        }
-                        delayedCommits.remove();
-                        Request request = new Request(p.hdr.getClientId(), p.hdr.getCxid(), p.hdr.getType(), p.hdr, p.rec, -1);
-                        request.setTxnDigest(p.digest);
-                        ozk.commitRequest(request);
-                    }
-                } else {
-                    // New server type need to handle in-flight packets
-                    throw new UnsupportedOperationException("Unknown server type");
-                }
-            }
-
-        }
-
-        SyncHelper helper = new SyncHelper();
-        QuorumPacket qp = new QuorumPacket();
+        // In the DIFF case we don't need to do a snapshot because the transactions will sync on top of any existing snapshot
+        // For SNAP and TRUNC the snapshot is needed to save that history
+        boolean snapshotNeeded = true;
+        boolean syncSnapshot = false;
         readPacket(qp);
+        Deque<Long> packetsCommitted = new ArrayDeque<>();
+        Deque<PacketInFlight> packetsNotLogged = new ArrayDeque<>();
+        Deque<PacketInFlight> packetsNotCommitted = new ArrayDeque<>();
         synchronized (zk) {
             if (qp.getType() == Leader.DIFF) {
                 LOG.info("Getting a diff from the leader 0x{}", Long.toHexString(qp.getZxid()));
                 self.setSyncMode(QuorumPeer.SyncMode.DIFF);
                 if (zk.shouldForceWriteInitialSnapshotAfterLeaderElection()) {
                     LOG.info("Forcing a snapshot write as part of upgrading from an older Zookeeper. This should only happen while upgrading.");
-                    helper.syncSnapshot();
+                    snapshotNeeded = true;
+                    syncSnapshot = true;
                 } else {
-                    helper.noSnapshot();
+                    snapshotNeeded = false;
                 }
             } else if (qp.getType() == Leader.SNAP) {
                 self.setSyncMode(QuorumPeer.SyncMode.SNAP);
                 LOG.info("Getting a snapshot from leader 0x{}", Long.toHexString(qp.getZxid()));
                 // The leader is going to dump the database
+                // db is clear as part of deserializeSnapshot()
                 zk.getZKDatabase().deserializeSnapshot(leaderIs);
                 // ZOOKEEPER-2819: overwrite config node content extracted
                 // from leader snapshot with local config, to avoid potential
@@ -692,18 +593,20 @@ public class Learner {
                 }
                 zk.getZKDatabase().setlastProcessedZxid(qp.getZxid());
 
-                // Immediately persist the latest snapshot when there is txn log gap
-                helper.syncSnapshot();
+                // immediately persist the latest snapshot when there is txn log gap
+                syncSnapshot = true;
             } else if (qp.getType() == Leader.TRUNC) {
-                // We need to truncate the log to the lastZxid of the leader
+                //we need to truncate the log to the lastzxid of the leader
                 self.setSyncMode(QuorumPeer.SyncMode.TRUNC);
                 LOG.warn("Truncating log to get in sync with the leader 0x{}", Long.toHexString(qp.getZxid()));
                 boolean truncated = zk.getZKDatabase().truncateLog(qp.getZxid());
                 if (!truncated) {
+                    // not able to truncate the log
                     LOG.error("Not able to truncate the log 0x{}", Long.toHexString(qp.getZxid()));
                     ServiceUtils.requestSystemExit(ExitCode.QUORUM_PACKET_ERROR.getValue());
                 }
                 zk.getZKDatabase().setlastProcessedZxid(qp.getZxid());
+
             } else {
                 LOG.error("Got unexpected packet from leader: {}, exiting ... ", LearnerHandler.packetToString(qp));
                 ServiceUtils.requestSystemExit(ExitCode.QUORUM_PACKET_ERROR.getValue());
@@ -711,10 +614,17 @@ public class Learner {
             zk.getZKDatabase().initConfigInZKDatabase(self.getQuorumVerifier());
             zk.createSessionTracker();
 
-
-            // we are now going to start getting transactions to apply followed by an UPTODATE
             long lastQueued = 0;
+
+            // in Zab V1.0 (ZK 3.4+) we might take a snapshot when we get the NEWLEADER message, but in pre V1.0
+            // we take the snapshot on the UPDATE message, since Zab V1.0 also gets the UPDATE (after the NEWLEADER)
+            // we need to make sure that we don't take the snapshot twice.
+            boolean isPreZAB1_0 = true;
+            //If we are not going to take the snapshot be sure the transactions are not applied in memory
+            // but written out to the transaction log
+            boolean writeToTxnLog = !snapshotNeeded;
             TxnLogEntry logEntry;
+            // we are now going to start getting transactions to apply followed by an UPTODATE
             outerLoop:
             while (self.isRunning()) {
                 readPacket(qp);
@@ -738,11 +648,13 @@ public class Learner {
                         QuorumVerifier qv = self.configFromString(new String(setDataTxn.getData(), UTF_8));
                         self.setLastSeenQuorumVerifier(qv, true);
                     }
-                    helper.propose(pif);
+
+                    packetsNotLogged.add(pif);
+                    packetsNotCommitted.add(pif);
                     break;
                 case Leader.COMMIT:
                 case Leader.COMMITANDACTIVATE:
-                    pif = helper.nextProposal();
+                    pif = packetsNotCommitted.peekFirst();
                     if (pif.hdr.getZxid() != qp.getZxid()) {
                         LOG.warn(
                             "Committing 0x{}, but next proposal is 0x{}",
@@ -750,24 +662,43 @@ public class Learner {
                             Long.toHexString(pif.hdr.getZxid()));
                     } else {
                         if (qp.getType() == Leader.COMMITANDACTIVATE) {
-                            tryReconfig(pif, ByteBuffer.wrap(qp.getData()).getLong(), qp.getZxid());
+                            QuorumVerifier qv = self.configFromString(new String(((SetDataTxn) pif.rec).getData(), UTF_8));
+                            boolean majorChange = self.processReconfig(
+                                qv,
+                                ByteBuffer.wrap(qp.getData()).getLong(), qp.getZxid(),
+                                true);
+                            if (majorChange) {
+                                throw new Exception("changes proposed in reconfig");
+                            }
                         }
-                        helper.commit();
+                        if (!writeToTxnLog) {
+                            zk.processTxn(pif.hdr, pif.rec);
+                            packetsNotLogged.remove();
+                            packetsNotCommitted.remove();
+                        } else {
+                            packetsNotCommitted.remove();
+                            packetsCommitted.add(qp.getZxid());
+                        }
                     }
                     break;
                 case Leader.INFORM:
                 case Leader.INFORMANDACTIVATE:
                     PacketInFlight packet = new PacketInFlight();
+
                     if (qp.getType() == Leader.INFORMANDACTIVATE) {
                         ByteBuffer buffer = ByteBuffer.wrap(qp.getData());
                         long suggestedLeaderId = buffer.getLong();
-                        byte[] remainingData = new byte[buffer.remaining()];
-                        buffer.get(remainingData);
-                        logEntry = SerializeUtils.deserializeTxn(remainingData);
+                        byte[] remainingdata = new byte[buffer.remaining()];
+                        buffer.get(remainingdata);
+                        logEntry = SerializeUtils.deserializeTxn(remainingdata);
                         packet.hdr = logEntry.getHeader();
                         packet.rec = logEntry.getTxn();
                         packet.digest = logEntry.getDigest();
-                        tryReconfig(packet, suggestedLeaderId, qp.getZxid());
+                        QuorumVerifier qv = self.configFromString(new String(((SetDataTxn) packet.rec).getData(), UTF_8));
+                        boolean majorChange = self.processReconfig(qv, suggestedLeaderId, qp.getZxid(), true);
+                        if (majorChange) {
+                            throw new Exception("changes proposed in reconfig");
+                        }
                     } else {
                         logEntry = SerializeUtils.deserializeTxn(qp.getData());
                         packet.rec = logEntry.getTxn();
@@ -782,8 +713,14 @@ public class Learner {
                         }
                         lastQueued = packet.hdr.getZxid();
                     }
-                    helper.propose(packet);
-                    helper.commit();
+                    if (!writeToTxnLog) {
+                        // Apply to db directly if we haven't taken the snapshot
+                        zk.processTxn(packet.hdr, packet.rec);
+                    } else {
+                        packetsNotLogged.add(packet);
+                        packetsCommitted.add(qp.getZxid());
+                    }
+
                     break;
                 case Leader.UPTODATE:
                     LOG.info("Learner received UPTODATE message");
@@ -793,11 +730,15 @@ public class Learner {
                             throw new Exception("changes proposed in reconfig");
                         }
                     }
-                    helper.flushAcks();
+                    if (isPreZAB1_0) {
+                        zk.takeSnapshot(syncSnapshot);
+                        self.setCurrentEpoch(newEpoch);
+                    }
                     self.setZooKeeperServer(zk);
                     self.adminServer.setZooKeeperServer(zk);
                     break outerLoop;
                 case Leader.NEWLEADER: // Getting NEWLEADER here instead of in discovery
+                    // means this is Zab 1.0
                     LOG.info("Learner received NEWLEADER message");
                     if (qp.getData() != null && qp.getData().length > 1) {
                         try {
@@ -809,13 +750,40 @@ public class Learner {
                         }
                     }
 
-                    helper.writeState();
+                    if (snapshotNeeded) {
+                        zk.takeSnapshot(syncSnapshot);
+                    }
+
+                    self.setCurrentEpoch(newEpoch);
+                    writeToTxnLog = true;
+                    //Anything after this needs to go to the transaction log, not applied directly in memory
+                    isPreZAB1_0 = false;
+
+                    // ZOOKEEPER-3911: make sure sync the uncommitted logs before commit them (ACK NEWLEADER).
+                    sock.setSoTimeout(self.tickTime * self.syncLimit);
+                    self.setSyncMode(QuorumPeer.SyncMode.NONE);
+                    zk.startupWithoutServing();
+                    if (zk instanceof FollowerZooKeeperServer) {
+                        FollowerZooKeeperServer fzk = (FollowerZooKeeperServer) zk;
+                        fzk.syncProcessor.setDelayForwarding(true);
+                        for (PacketInFlight p : packetsNotLogged) {
+                            fzk.logRequest(p.hdr, p.rec, p.digest);
+                        }
+                        packetsNotLogged.clear();
+                        fzk.syncProcessor.syncFlush();
+                    }
+
                     writePacket(new QuorumPacket(Leader.ACK, newLeaderZxid, null, null), true);
+
+                    if (zk instanceof FollowerZooKeeperServer) {
+                        FollowerZooKeeperServer fzk = (FollowerZooKeeperServer) zk;
+                        fzk.syncProcessor.setDelayForwarding(false);
+                        fzk.syncProcessor.syncFlush();
+                    }
                     break;
                 }
             }
         }
-        QuorumPacket ack = new QuorumPacket(Leader.ACK, 0, null, null);
         ack.setZxid(ZxidUtils.makeZxid(newEpoch, 0));
         writePacket(ack, true);
         zk.startServing();
@@ -828,14 +796,40 @@ public class Learner {
          */
         self.updateElectionVote(newEpoch);
 
-        helper.applyDelayedPackets();
-    }
-
-    private void tryReconfig(PacketInFlight pif, long newLeader, long zxid) throws Exception {
-        QuorumVerifier qv = self.configFromString(new String(((SetDataTxn) pif.rec).getData(), UTF_8));
-        boolean majorChange = self.processReconfig(qv, newLeader, zxid, true);
-        if (majorChange) {
-            throw new Exception("changes proposed in reconfig");
+        // We need to log the stuff that came in between the snapshot and the uptodate
+        if (zk instanceof FollowerZooKeeperServer) {
+            FollowerZooKeeperServer fzk = (FollowerZooKeeperServer) zk;
+            for (PacketInFlight p : packetsNotLogged) {
+                fzk.logRequest(p.hdr, p.rec, p.digest);
+            }
+            for (Long zxid : packetsCommitted) {
+                fzk.commit(zxid);
+            }
+        } else if (zk instanceof ObserverZooKeeperServer) {
+            // Similar to follower, we need to log requests between the snapshot
+            // and UPTODATE
+            ObserverZooKeeperServer ozk = (ObserverZooKeeperServer) zk;
+            for (PacketInFlight p : packetsNotLogged) {
+                Long zxid = packetsCommitted.peekFirst();
+                if (p.hdr.getZxid() != zxid) {
+                    // log warning message if there is no matching commit
+                    // old leader send outstanding proposal to observer
+                    LOG.warn(
+                        "Committing 0x{}, but next proposal is 0x{}",
+                        Long.toHexString(zxid),
+                        Long.toHexString(p.hdr.getZxid()));
+                    continue;
+                }
+                packetsCommitted.remove();
+                Request request = new Request(null, p.hdr.getClientId(), p.hdr.getCxid(), p.hdr.getType(), null, null);
+                request.setTxn(p.rec);
+                request.setHdr(p.hdr);
+                request.setTxnDigest(p.digest);
+                ozk.commitRequest(request);
+            }
+        } else {
+            // New server type need to handle in-flight packets
+            throw new UnsupportedOperationException("Unknown server type");
         }
     }
