--- zookeeper-server/zookeeper-server-3.9.1/CMakeLists.txt
@@ -1,2 +1,0 @@
-# Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-install_jar(zookeeper-server-3.9.1-jar-with-dependencies.jar)
