+++ zookeeper-server/zookeeper-server/CMakeLists.txt
@@ -1,4 +1,2 @@
 # Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-install_jar(zookeeper-server-3.8.0-jar-with-dependencies.jar)
-# Make symlink so that we have a default version, should be done only in zookeeper-server module
-install_symlink(lib/jars/zookeeper-server-3.8.0-jar-with-dependencies.jar lib/jars/zookeeper-server-jar-with-dependencies.jar)
+install_jar(zookeeper-server-3.9.1-jar-with-dependencies.jar)
